UI - Active
액티브 인벤까지 완료
아이템 사용하면 쿨타임 줄어요 다시 차는거는 아직 미구현

+++ Assets/Items/Scripts/ScriptableObject/ActiveItemData.cs
@@ -27,5 +27,5 @@ public class ActiveItemData : ItemData {
     /// <summary>
     /// 아이템 쿨타임
     /// </summary>
-    public float coolTime;
+    public int coolTime;
 }
